% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Visualize FARS Accidents by State and Year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a numeric vector of length 1}

\item{year}{A numeric vector of length 1}
}
\value{
a plot object
}
\description{
This function takes year and state.num inputs and plots a map of the
selected state with all accidents from that state and year.
}
\note{
This function will throw an error if the state number given is not an
actual state number.

This function will return a NULL object if there are no accidents to
plot.
}
\examples{
\dontrun{
fars_map_state(1, 2013)
fars_map_state(30, 2015)
}

}
