% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string providing the filename for the function
to read in.}
}
\value{
A dplyr tbl_df
}
\description{
This function reads in a given file and turns into a table.
}
\note{
the function will thrown an error if the filename provided as input
does not exist
}
\examples{
\dontrun{
fars_read("~/inputs/fars1.csv")
fars_read("~/in/fars2.csv.bz2")
}

}
