% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Multiple Years of FARS Data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a numeric vector of length >= 1}
}
\value{
A list of tables
}
\description{
This function takes a vector of years as input and returns a
list of tables containing the month and year variables from the FARS data
tables. The years returned correspond to the years in the input vector.
}
\note{
The function will throw an 'invalid year' error if it encounters any
error.
}
\examples{
\dontrun{
fars_read_years(c(2000, 2002))
fars_read-years(2013:2015)
}

}
