% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Multiple Years of FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a numeric vector of length >= 1}
}
\value{
a table with size 12 x (length(years)) + 1
}
\description{
This function takes a list of years and returns a table with a count
of the number of accidents per month for each year
}
\examples{
\dontrun{
fars_summarize_years(c(2000, 2002))
fars_summarize_years(2012:2015)
}

}
